+++ tuk-api/src/test/kotlin/nexters/tuk/application/gathering/GatheringServiceIntegrationTest.kt
@@ -13,9 +13,11 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
+import org.junit.jupiter.api.assertNull
 import org.junit.jupiter.api.assertThrows
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.data.repository.findByIdOrNull
 
 @SpringBootTest
 class GatheringServiceIntegrationTest @Autowired constructor(
@@ -211,6 +213,7 @@ class GatheringServiceIntegrationTest @Autowired constructor(
         // when
         val actual = gatheringService.updateGathering(
             GatheringCommand.Update(
+                memberId = 1L,
                 gatheringId = gathering.id,
                 gatheringIntervalDays = 30
             )
@@ -224,5 +227,78 @@ class GatheringServiceIntegrationTest @Autowired constructor(
                     .isEqualTo(30)
             },
         )
+    }
+
+    @Test
+    fun `호스트가 아닌 사용자는 모임을 수정할 수 없다`() {
+        // given
+        val gathering = gatheringRepository.save(
+            Gathering.generate(
+                hostId = 1L,
+                name = "모임",
+                intervalDays = 10
+            )
+        )
+
+        // when & then
+        val exception = assertThrows<BaseException> {
+            gatheringService.updateGathering(
+                GatheringCommand.Update(
+                    memberId = 2L,
+                    gatheringId = gathering.id,
+                    gatheringIntervalDays = 30
+                )
+            )
+        }
+
+        assertThat(exception.message).isEqualTo("수정 권한이 없습니다.")
+    }
+
+    @Test
+    fun `호스트가 모임을 삭제할 수 있다`() {
+        // given
+        val gathering = gatheringRepository.save(
+            Gathering.generate(
+                hostId = 1L,
+                name = "모임",
+                intervalDays = 10
+            )
+        )
+
+        // when
+        gatheringService.deleteGathering(
+            GatheringCommand.Delete(
+                memberId = 1L,
+                gatheringId = gathering.id
+            )
+        )
+
+        // then
+        val deletedGathering = gatheringRepository.findByIdOrNull(gathering.id)
+        assertNull(deletedGathering)
+    }
+
+    @Test
+    fun `호스트가 아닌 사용자는 모임을 삭제할 수 없다`() {
+        // given
+        val gathering = gatheringRepository.save(
+            Gathering.generate(
+                hostId = 1L,
+                name = "모임",
+                intervalDays = 10
+            )
+        )
+
+        // when & then
+        val exception = assertThrows<BaseException> {
+            gatheringService.deleteGathering(
+                GatheringCommand.Delete(
+                    memberId = 2L,
+                    gatheringId = gathering.id
+                )
+            )
+        }
+
+        assertThat(exception.message).isEqualTo("수정 권한이 없습니다.")
     }
 }
